feat/rename-project: Renomeando o nome do projeto
