+++ app/src/androidTest/java/com/example/igestamobile/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.igestamobile;
 
 import android.content.Context;
