+++ app/src/test/java/com/example/igestamobile/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.igestamobile;
 
 import org.junit.Test;
